Move honk/pcs/ to commitment_schemes/
