+++ barretenberg/cpp/src/barretenberg/commitment_schemes/ipa/ipa.test.cpp
@@ -1,73 +1,97 @@
-
-#include "kzg.hpp"
+#include "ipa.hpp"
 #include "../gemini/gemini.hpp"
 #include "../shplonk/shplonk.hpp"
-
-#include "../commitment_key.test.hpp"
-#include "barretenberg/honk/pcs/claim.hpp"
+#include "barretenberg/commitment_schemes/commitment_key.test.hpp"
+#include "barretenberg/common/mem.hpp"
+#include "barretenberg/ecc/curves/bn254/fq12.hpp"
+#include "barretenberg/ecc/curves/types.hpp"
 #include "barretenberg/polynomials/polynomial.hpp"
-
-#include "barretenberg/ecc/curves/bn254/g1.hpp"
-
+#include "barretenberg/polynomials/polynomial_arithmetic.hpp"
 #include <gtest/gtest.h>
-#include <vector>
+using namespace barretenberg;
+namespace proof_system::honk::pcs::ipa::test {
 
-namespace proof_system::honk::pcs::kzg {
+using Curve = curve::Grumpkin;
 
-template <class Curve> class KZGTest : public CommitmentTest<Curve> {
+class IPATest : public CommitmentTest<Curve> {
   public:
     using Fr = typename Curve::ScalarField;
-    using Commitment = typename Curve::AffineElement;
     using GroupElement = typename Curve::Element;
+    using CK = CommitmentKey<Curve>;
+    using VK = VerifierCommitmentKey<Curve>;
     using Polynomial = barretenberg::Polynomial<Fr>;
 };
 
-TYPED_TEST_SUITE(KZGTest, CommitmentSchemeParams);
-
-TYPED_TEST(KZGTest, single)
+TEST_F(IPATest, CommitOnManyZeroCoeffPolyWorks)
 {
-    const size_t n = 16;
-
-    using KZG = KZG<TypeParam>;
-    using Fr = typename TypeParam::ScalarField;
-
-    auto witness = this->random_polynomial(n);
-    barretenberg::g1::element commitment = this->commit(witness);
-
-    auto challenge = Fr::random_element();
-    auto evaluation = witness.evaluate(challenge);
-    auto opening_pair = OpeningPair<TypeParam>{ challenge, evaluation };
-    auto opening_claim = OpeningClaim<TypeParam>{ opening_pair, commitment };
-
-    auto prover_transcript = BaseTranscript<Fr>::prover_init_empty();
-
-    KZG::compute_opening_proof(this->ck(), opening_pair, witness, prover_transcript);
-
-    auto verifier_transcript = BaseTranscript<Fr>::verifier_init_empty(prover_transcript);
-    bool verified = KZG::verify(this->vk(), opening_claim, verifier_transcript);
-
-    EXPECT_EQ(verified, true);
+    constexpr size_t n = 4;
+    Polynomial p(n);
+    for (size_t i = 0; i < n - 1; i++) {
+        p[i] = Fr::zero();
+    }
+    p[3] = Fr::one();
+    GroupElement commitment = this->commit(p);
+    auto srs_elements = this->ck()->srs->get_monomial_points();
+    GroupElement expected = srs_elements[0] * p[0];
+    // The SRS stored in the commitment key is the result after applying the pippenger point table so the
+    // values at odd indices contain the point {srs[i-1].x * beta, srs[i-1].y}, where beta is the endomorphism
+    // G_vec_local should use only the original SRS thus we extract only the even indices.
+    for (size_t i = 2; i < 2 * n; i += 2) {
+        expected += srs_elements[i] * p[i >> 1];
+    }
+    EXPECT_EQ(expected.normalize(), commitment.normalize());
 }
 
-/**
- * @brief Test full PCS protocol: Gemini, Shplonk, KZG and pairing check
- * @details Demonstrates the full PCS protocol as it is used in the construction and verification
- * of a single Honk proof. (Expository comments included throughout).
- *
- */
-TYPED_TEST(KZGTest, GeminiShplonkKzgWithShift)
+TEST_F(IPATest, Commit)
 {
-    using ShplonkProver = shplonk::ShplonkProver_<TypeParam>;
-    using ShplonkVerifier = shplonk::ShplonkVerifier_<TypeParam>;
-    using GeminiProver = gemini::GeminiProver_<TypeParam>;
-    using GeminiVerifier = gemini::GeminiVerifier_<TypeParam>;
-    using KZG = KZG<TypeParam>;
-    using Fr = typename TypeParam::ScalarField;
-    using GroupElement = typename TypeParam::Element;
-    using Polynomial = typename barretenberg::Polynomial<Fr>;
+    constexpr size_t n = 128;
+    auto poly = this->random_polynomial(n);
+    GroupElement commitment = this->commit(poly);
+    auto srs_elements = this->ck()->srs->get_monomial_points();
+    GroupElement expected = srs_elements[0] * poly[0];
+    // The SRS stored in the commitment key is the result after applying the pippenger point table so the
+    // values at odd indices contain the point {srs[i-1].x * beta, srs[i-1].y}, where beta is the endomorphism
+    // G_vec_local should use only the original SRS thus we extract only the even indices.
+    for (size_t i = 2; i < 2 * n; i += 2) {
+        expected += srs_elements[i] * poly[i >> 1];
+    }
+    EXPECT_EQ(expected.normalize(), commitment.normalize());
+}
 
-    const size_t n = 16;
-    const size_t log_n = 4;
+TEST_F(IPATest, Open)
+{
+    using IPA = IPA<Curve>;
+    // generate a random polynomial, degree needs to be a power of two
+    size_t n = 128;
+    auto poly = this->random_polynomial(n);
+    auto [x, eval] = this->random_eval(poly);
+    auto commitment = this->commit(poly);
+    const OpeningPair<Curve> opening_pair = { x, eval };
+    const OpeningClaim<Curve> opening_claim{ opening_pair, commitment };
+
+    // initialize empty prover transcript
+    BaseTranscript<Fr> prover_transcript;
+    IPA::compute_opening_proof(this->ck(), opening_pair, poly, prover_transcript);
+
+    // initialize verifier transcript from proof data
+    BaseTranscript<Fr> verifier_transcript{ prover_transcript.proof_data };
+
+    auto result = IPA::verify(this->vk(), opening_claim, verifier_transcript);
+    EXPECT_TRUE(result);
+
+    EXPECT_EQ(prover_transcript.get_manifest(), verifier_transcript.get_manifest());
+}
+
+TEST_F(IPATest, GeminiShplonkIPAWithShift)
+{
+    using IPA = IPA<Curve>;
+    using ShplonkProver = shplonk::ShplonkProver_<Curve>;
+    using ShplonkVerifier = shplonk::ShplonkVerifier_<Curve>;
+    using GeminiProver = gemini::GeminiProver_<Curve>;
+    using GeminiVerifier = gemini::GeminiVerifier_<Curve>;
+
+    const size_t n = 8;
+    const size_t log_n = 3;
 
     Fr rho = Fr::random_element();
 
@@ -85,25 +109,21 @@ TYPED_TEST(KZGTest, GeminiShplonkKzgWithShift)
     auto eval2 = poly2.evaluate_mle(mle_opening_point);
     auto eval2_shift = poly2.evaluate_mle(mle_opening_point, true);
 
-    // Collect multilinear evaluations for input to prover
     std::vector<Fr> multilinear_evaluations = { eval1, eval2, eval2_shift };
 
     std::vector<Fr> rhos = gemini::powers_of_rho(rho, multilinear_evaluations.size());
 
-    // Compute batched multivariate evaluation
     Fr batched_evaluation = Fr::zero();
     for (size_t i = 0; i < rhos.size(); ++i) {
         batched_evaluation += multilinear_evaluations[i] * rhos[i];
     }
 
-    // Compute batched polynomials
     Polynomial batched_unshifted(n);
     Polynomial batched_to_be_shifted(n);
     batched_unshifted.add_scaled(poly1, rhos[0]);
     batched_unshifted.add_scaled(poly2, rhos[1]);
     batched_to_be_shifted.add_scaled(poly2, rhos[2]);
 
-    // Compute batched commitments
     GroupElement batched_commitment_unshifted = GroupElement::zero();
     GroupElement batched_commitment_to_be_shifted = GroupElement::zero();
     batched_commitment_unshifted = commitment1 * rhos[0] + commitment2 * rhos[1];
@@ -111,11 +131,6 @@ TYPED_TEST(KZGTest, GeminiShplonkKzgWithShift)
 
     auto prover_transcript = BaseTranscript<Fr>::prover_init_empty();
 
-    // Run the full prover PCS protocol:
-
-    // Compute:
-    // - (d+1) opening pairs: {r, \hat{a}_0}, {-r^{2^i}, a_i}, i = 0, ..., d-1
-    // - (d+1) Fold polynomials Fold_{r}^(0), Fold_{-r}^(0), and Fold^(i), i = 0, ..., d-1
     auto gemini_polynomials = GeminiProver::compute_gemini_polynomials(
         mle_opening_point, std::move(batched_unshifted), std::move(batched_to_be_shifted));
 
@@ -136,9 +151,6 @@ TYPED_TEST(KZGTest, GeminiShplonkKzgWithShift)
         prover_transcript.send_to_verifier(label, evaluation);
     }
 
-    // Shplonk prover output:
-    // - opening pair: (z_challenge, 0)
-    // - witness: polynomial Q - Q_z
     const Fr nu_challenge = prover_transcript.get_challenge("Shplonk:nu");
     auto batched_quotient_Q =
         ShplonkProver::compute_batched_quotient(gemini_opening_pairs, gemini_witnesses, nu_challenge);
@@ -148,33 +160,20 @@ TYPED_TEST(KZGTest, GeminiShplonkKzgWithShift)
     const auto [shplonk_opening_pair, shplonk_witness] = ShplonkProver::compute_partially_evaluated_batched_quotient(
         gemini_opening_pairs, gemini_witnesses, std::move(batched_quotient_Q), nu_challenge, z_challenge);
 
-    // KZG prover:
-    // - Adds commitment [W] to transcript
-    KZG::compute_opening_proof(this->ck(), shplonk_opening_pair, shplonk_witness, prover_transcript);
-
-    // Run the full verifier PCS protocol with genuine opening claims (genuine commitment, genuine evaluation)
+    IPA::compute_opening_proof(this->ck(), shplonk_opening_pair, shplonk_witness, prover_transcript);
 
     auto verifier_transcript = BaseTranscript<Fr>::verifier_init_empty(prover_transcript);
 
-    // Gemini verifier output:
-    // - claim: d+1 commitments to Fold_{r}^(0), Fold_{-r}^(0), Fold^(l), d+1 evaluations a_0_pos, a_l, l = 0:d-1
     auto gemini_verifier_claim = GeminiVerifier::reduce_verification(mle_opening_point,
                                                                      batched_evaluation,
                                                                      batched_commitment_unshifted,
                                                                      batched_commitment_to_be_shifted,
                                                                      verifier_transcript);
 
-    // Shplonk verifier claim: commitment [Q] - [Q_z], opening point (z_challenge, 0)
     const auto shplonk_verifier_claim =
         ShplonkVerifier::reduce_verification(this->vk(), gemini_verifier_claim, verifier_transcript);
-
-    // KZG verifier:
-    // aggregates inputs [Q] - [Q_z] and [W] into an 'accumulator' (can perform pairing check on result)
-    bool verified = KZG::verify(this->vk(), shplonk_verifier_claim, verifier_transcript);
-
-    // Final pairing check: e([Q] - [Q_z] + z[W], [1]_2) = e([W], [x]_2)
+    bool verified = IPA::verify(this->vk(), shplonk_verifier_claim, verifier_transcript);
 
     EXPECT_EQ(verified, true);
 }
-
-} // namespace proof_system::honk::pcs::kzg
+} // namespace proof_system::honk::pcs::ipa::test
