+++ barretenberg/cpp/src/barretenberg/honk/proof_system/goblin_merge/merge_verifier.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
+#include "barretenberg/commitment_schemes/claim.hpp"
 #include "barretenberg/honk/flavor/goblin_ultra.hpp"
 #include "barretenberg/honk/flavor/ultra.hpp"
-#include "barretenberg/honk/pcs/claim.hpp"
 #include "barretenberg/plonk/proof_system/types/proof.hpp"
 #include "barretenberg/proof_system/op_queue/ecc_op_queue.hpp"
 #include "barretenberg/transcript/transcript.hpp"
