+++ barretenberg/cpp/src/barretenberg/honk/proof_system/ultra_verifier.cpp
@@ -1,5 +1,5 @@
 #include "./ultra_verifier.hpp"
-#include "barretenberg/honk/pcs/zeromorph/zeromorph.hpp"
+#include "barretenberg/commitment_schemes/zeromorph/zeromorph.hpp"
 #include "barretenberg/honk/utils/power_polynomial.hpp"
 #include "barretenberg/numeric/bitop/get_msb.hpp"
 #include "barretenberg/transcript/transcript.hpp"
